requirements.txt
annotated-types==0.7.0
anyio==4.11.0
arrow==1.3.0
attrs==25.3.0
Authlib==1.6.4
boolean.py==5.0
certifi==2025.8.3
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==46.0.1
cyclonedx-bom==7.1.0
cyclonedx-python-lib==11.1.0
dataclasses-json==0.6.7
defusedxml==0.7.1
dparse==0.6.4
fastapi==0.118.0
filelock==3.19.1
fqdn==1.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
isoduration==20.11.0
Jinja2==3.1.6
joblib==1.5.2
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lark==1.3.0
license-expression==30.4.4
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
mdurl==0.1.2
mypy_extensions==1.1.0
nltk==3.9.1
packageurl-python==0.17.5
packaging==25.0
pip-requirements-parser==32.0.1
pluggy==1.6.0
psutil==7.1.0
py-serializable==2.1.0
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.5
pytest==8.4.2
python-dateutil==2.9.0.post0
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
rpds-py==0.27.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
ruff==0.13.2
safety==3.6.2
safety-schemas==0.0.16
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.48.0
tenacity==9.1.2
tomlkit==0.13.3
tqdm==4.67.1
typer==0.19.2
types-python-dateutil==2.9.0.20250822
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.37.0
webcolors==24.11.1
python-multipart

python-multipart>=0.0.9
prometheus-client>=0.20.0
